requirements.txt
# Web Framework
Flask>=2.3.0
Flask-Cors>=4.0.0

# Machine Learning and Data Processing
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0
imbalanced-learn>=0.11.0

# Environment Variables
python-dotenv>=1.0.0

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# HTTP Requests
requests>=2.31.0

# Runtime Dependencies
Werkzeug>=3.0.0
click>=8.1.7
itsdangerous>=2.1.2
Jinja2>=3.1.2
MarkupSafe>=2.1.3